Core Layer Added

+++ Entities/Concrete/Category.cs
@@ -3,13 +3,13 @@ using System.Collections.Generic;
 using System.Linq;
 using System.Text;
 using System.Threading.Tasks;
-using Entities.Abstract;
+using Core.Entities;
 
 namespace Entities.Concrete
 {
     public class Category : IEntity
     {
-        public int Id { get; set; }
+        public byte Id { get; set; }
         public string CategoryName1 { get; set; }
         public string CategoryName2 { get; set; }
         public string CategoryName3 { get; set; }
